migrate migrations; split fields into `fields.py`
